Revert "Overlays"
