+++ MapKit_Dimo/Assets.xcassets/AppIcon.appiconset/Contents.json
