+++ MapKit_Dimo/Assets.xcassets/Contents.json
